This commit add the ability to create/delete VMs
* Defaults file with some defaults and a HW profile
* vars file for input validation
* created extra defaults files for existing actions to ensure
  everything uses the same mechanism
* Added task to load defaults alongside vars

TODO: docs

NOTE: As of this time, the first attempt to create the machine seems to
fail, because the customization can't be run. This also occurs when
there is no customization to run. This seems like a known bug with no
released fix as of now. To work around this, the task will simply retry
2 times before actually failing. This has fixed at least any issues I
ran into in my lab testing.